Remove claims chat
