--- apollo/src/main/graphql/com/hedvig/android/owldroid/graphql/triggerClaimChat.graphql
@@ -1,3 +1,0 @@
-mutation TriggerClaimChat($input: TriggerClaimChatInput!) {
-  triggerClaimChat(input: $input)
-}
