+++ app/src/main/java/com/hedvig/app/feature/claims/data/ClaimsRepository.kt
@@ -1,11 +1,8 @@
 package com.hedvig.app.feature.claims.data
 
 import com.apollographql.apollo.ApolloClient
-import com.apollographql.apollo.api.Input
 import com.apollographql.apollo.coroutines.await
 import com.hedvig.android.owldroid.graphql.CommonClaimQuery
-import com.hedvig.android.owldroid.graphql.TriggerClaimChatMutation
-import com.hedvig.android.owldroid.type.TriggerClaimChatInput
 import com.hedvig.app.util.LocaleManager
 
 class ClaimsRepository(
@@ -15,11 +12,4 @@ class ClaimsRepository(
 
     suspend fun fetchCommonClaims() = apolloClient
         .query(CommonClaimQuery(localeManager.defaultLocale())).await()
-
-    suspend fun triggerClaimsChat(claimTypeId: String?) = apolloClient
-        .mutate(
-            TriggerClaimChatMutation(
-                TriggerClaimChatInput(Input.fromNullable(claimTypeId))
-            )
-        ).await()
 }
