+++ app/src/main/java/com/hedvig/app/feature/claims/ui/ClaimsViewModel.kt
@@ -3,10 +3,13 @@ package com.hedvig.app.feature.claims.ui
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import arrow.core.Either
 import com.hedvig.android.owldroid.graphql.CommonClaimQuery
 import com.hedvig.app.feature.chat.data.ChatRepository
 import com.hedvig.app.feature.claims.data.ClaimsRepository
 import e
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
 
 class ClaimsViewModel(
@@ -15,6 +18,14 @@ class ClaimsViewModel(
 ) : ViewModel() {
 
     val data: MutableLiveData<CommonClaimQuery.Data> = MutableLiveData()
+
+    sealed class Event {
+        object StartChat : Event()
+        object Error : Event()
+    }
+
+    private val _events = Channel<Event>(Channel.UNLIMITED)
+    val events = _events.receiveAsFlow()
 
     init {
         fetchCommonClaims()
@@ -27,27 +38,17 @@ class ClaimsViewModel(
                 response.exceptionOrNull()?.let { e { "$it Failed to fetch claims data" } }
                 return@launch
             }
-            data.postValue(response.getOrNull()?.data)
-        }
-    }
-
-    suspend fun triggerClaimsChat(claimTypeId: String? = null) {
-        val response =
-            runCatching { claimsRepository.triggerClaimsChat(claimTypeId) }
-        if (response.isFailure) {
-            response.exceptionOrNull()?.let { e(it) }
-            return
+            response.getOrNull()?.data?.let(data::postValue)
         }
     }
 
     suspend fun triggerFreeTextChat() {
-        val response = runCatching {
-            chatRepository
-                .triggerFreeTextChat()
-        }
-        if (response.isFailure) {
-            response.exceptionOrNull()?.let { e(it) }
-            return
+        viewModelScope.launch {
+            val event = when (val result = chatRepository.triggerFreeTextChat()) {
+                is Either.Left -> Event.Error
+                is Either.Right -> Event.StartChat
+            }
+            _events.trySend(event)
         }
     }
 }
