+++ app/src/main/java/com/hedvig/app/feature/claims/ui/pledge/HonestyPledgeBottomSheet.kt
@@ -9,22 +9,17 @@ import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.hedvig.app.R
 import com.hedvig.app.databinding.BottomSheetHonestyPledgeBinding
 import com.hedvig.app.feature.claims.service.ClaimsTracker
-import com.hedvig.app.feature.claims.ui.ClaimsViewModel
 import com.hedvig.app.feature.embark.ui.EmbarkActivity
 import com.hedvig.app.util.extensions.view.setHapticClickListener
 import com.hedvig.app.util.extensions.viewLifecycleScope
-import com.hedvig.app.util.featureflags.FeatureManager
 import com.zhuinden.fragmentviewbindingdelegatekt.viewBinding
 import org.koin.android.ext.android.inject
-import org.koin.androidx.viewmodel.ext.android.sharedViewModel
 
 class HonestyPledgeBottomSheet(
     private val customActivityLaunch: ((Intent) -> Unit)? = null,
 ) : BottomSheetDialogFragment() {
     private val tracker: ClaimsTracker by inject()
-    private val claimsViewModel: ClaimsViewModel by sharedViewModel()
     private val binding by viewBinding(BottomSheetHonestyPledgeBinding::bind)
-    private val featureManager: FeatureManager by inject()
 
     override fun onCreateView(
         inflater: LayoutInflater,
