+++ app/src/main/java/com/hedvig/app/feature/offer/OfferViewModel.kt
@@ -11,7 +11,7 @@ import com.hedvig.android.owldroid.type.SignMethod
 import com.hedvig.app.authenticate.LoginStatus
 import com.hedvig.app.authenticate.LoginStatusService
 import com.hedvig.app.feature.adyen.AdyenRepository
-import com.hedvig.app.feature.chat.usecase.TriggerFreeTextChatUseCase
+import com.hedvig.app.feature.chat.data.ChatRepository
 import com.hedvig.app.feature.documents.DocumentItems
 import com.hedvig.app.feature.insurablelimits.InsurableLimitItem
 import com.hedvig.app.feature.offer.quotedetail.buildDocuments
@@ -153,7 +153,7 @@ class OfferViewModelImpl(
     private val tracker: OfferTracker,
     private val adyenRepository: AdyenRepository,
     private val marketManager: MarketManager,
-    private val freeTextChatUseCase: TriggerFreeTextChatUseCase
+    private val chatRepository: ChatRepository,
 ) : OfferViewModel() {
 
     private lateinit var quoteIds: List<String>
@@ -366,7 +366,7 @@ class OfferViewModelImpl(
         offerRepository.writeDiscountToCache(quoteIds, data)
 
     override suspend fun triggerOpenChat() {
-        val event = when (freeTextChatUseCase.invoke()) {
+        val event = when (chatRepository.triggerFreeTextChat()) {
             is Either.Left -> Event.Error
             is Either.Right -> Event.OpenChat
         }
